Updated winui chart sample with latest project reunion.
