+++ ChartGettingStartedDesktop/ChartGettingStartedDesktop.sln
@@ -1,0 +1,57 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31025.218
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{C7167F0D-BC9F-4E6E-AFE1-012C56B48DB5}") = "ChartGettingStartedDesktop (Package)", "ChartGettingStartedDesktop (Package)\ChartGettingStartedDesktop (Package).wapproj", "{D70BBD51-F981-477F-9C31-4B2BF494290D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChartGettingStartedDesktop", "ChartGettingStartedDesktop\ChartGettingStartedDesktop.csproj", "{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|arm64 = Debug|arm64
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|arm64 = Release|arm64
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Debug|arm64.ActiveCfg = Debug|arm64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Debug|arm64.Build.0 = Debug|arm64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Debug|arm64.Deploy.0 = Debug|arm64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Debug|x64.ActiveCfg = Debug|x64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Debug|x64.Build.0 = Debug|x64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Debug|x64.Deploy.0 = Debug|x64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Debug|x86.ActiveCfg = Debug|x86
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Debug|x86.Build.0 = Debug|x86
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Debug|x86.Deploy.0 = Debug|x86
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Release|arm64.ActiveCfg = Release|arm64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Release|arm64.Build.0 = Release|arm64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Release|arm64.Deploy.0 = Release|arm64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Release|x64.ActiveCfg = Release|x64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Release|x64.Build.0 = Release|x64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Release|x64.Deploy.0 = Release|x64
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Release|x86.ActiveCfg = Release|x86
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Release|x86.Build.0 = Release|x86
+		{D70BBD51-F981-477F-9C31-4B2BF494290D}.Release|x86.Deploy.0 = Release|x86
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Debug|arm64.ActiveCfg = Debug|arm64
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Debug|arm64.Build.0 = Debug|arm64
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Debug|x64.ActiveCfg = Debug|x64
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Debug|x64.Build.0 = Debug|x64
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Debug|x86.ActiveCfg = Debug|x86
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Debug|x86.Build.0 = Debug|x86
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Release|arm64.ActiveCfg = Release|arm64
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Release|arm64.Build.0 = Release|arm64
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Release|x64.ActiveCfg = Release|x64
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Release|x64.Build.0 = Release|x64
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Release|x86.ActiveCfg = Release|x86
+		{E51522D8-1420-42D1-AE31-71B0F4FB5AC3}.Release|x86.Build.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {DC20B4CA-FEF3-4CAC-98F3-6403180A96EE}
+	EndGlobalSection
+EndGlobal
