+++ ChartGettingStartedDesktop/ChartGettingStartedDesktop/Model/Model.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ChartGettingStartedDesktop
+{
+    public class Person
+    {
+        public string Name { get; set; }
+
+        public double Height { get; set; }
+    }
+}
