+++ ChartGettingStartedUWP/ChartGettingStartedUWP.sln
@@ -1,0 +1,43 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31025.218
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ChartGettingStartedUWP", "ChartGettingStartedUWP.csproj", "{D0D55F54-DCC0-4A56-AA42-8EC91E734488}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|arm64 = Debug|arm64
+		Debug|x64 = Debug|x64
+		Debug|x86 = Debug|x86
+		Release|arm64 = Release|arm64
+		Release|x64 = Release|x64
+		Release|x86 = Release|x86
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Debug|arm64.ActiveCfg = Debug|arm64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Debug|arm64.Build.0 = Debug|arm64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Debug|arm64.Deploy.0 = Debug|arm64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Debug|x64.ActiveCfg = Debug|x64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Debug|x64.Build.0 = Debug|x64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Debug|x64.Deploy.0 = Debug|x64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Debug|x86.ActiveCfg = Debug|x86
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Debug|x86.Build.0 = Debug|x86
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Debug|x86.Deploy.0 = Debug|x86
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Release|arm64.ActiveCfg = Release|arm64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Release|arm64.Build.0 = Release|arm64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Release|arm64.Deploy.0 = Release|arm64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Release|x64.ActiveCfg = Release|x64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Release|x64.Build.0 = Release|x64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Release|x64.Deploy.0 = Release|x64
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Release|x86.ActiveCfg = Release|x86
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Release|x86.Build.0 = Release|x86
+		{D0D55F54-DCC0-4A56-AA42-8EC91E734488}.Release|x86.Deploy.0 = Release|x86
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {290FDC0F-9609-4FA9-89BD-388E753CF6C7}
+	EndGlobalSection
+EndGlobal
