+++ ChartGettingStartedUWP/MainPage.xaml
@@ -1,0 +1,50 @@
+﻿<Page
+    x:Class="ChartGettingStartedUWP.MainPage"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:local="using:ChartGettingStartedUWP"
+    xmlns:chart="using:Syncfusion.UI.Xaml.Charts"
+    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+    mc:Ignorable="d"
+    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}">
+
+    <Grid>
+
+        <Grid.DataContext>
+            <local:ViewModel/>
+        </Grid.DataContext>
+
+        <chart:SfChart Header="Chart">
+            
+            <!--Initialize the horizontal axis for SfChart-->
+            <chart:SfChart.PrimaryAxis>
+                <chart:CategoryAxis Header="Name" />
+            </chart:SfChart.PrimaryAxis>
+
+            <chart:SfChart.SecondaryAxis>
+                <chart:NumericalAxis Header="Height(in cm)"/>
+            </chart:SfChart.SecondaryAxis>
+
+            <!--Adding Legend to the SfChart-->
+            <chart:SfChart.Legend>
+                <chart:ChartLegend />
+            </chart:SfChart.Legend>
+
+            <!--Initialize the series for SfChart-->
+            <chart:ColumnSeries  ItemsSource="{Binding Data}"
+                                      XBindingPath="Name"
+                                      YBindingPath="Height"
+                                      Label="Heights"
+                                      ShowTooltip="True">
+                <!--Add data marker to the series-->
+                <chart:ColumnSeries.DataMarker>
+                    <chart:ChartDataMarker DataMarkerPosition="Top" LabelPosition="Inner"  ShowLabel="True" />
+                </chart:ColumnSeries.DataMarker>
+            </chart:ColumnSeries>
+
+        </chart:SfChart>
+
+    </Grid>
+
+</Page>
