+++ ChartGettingStartedUWP/ViewModel/ViewModel.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Collections.ObjectModel;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace ChartGettingStartedUWP
+{
+    public class ViewModel
+    {
+        public ObservableCollection<Person> Data { get; set; }
+
+        public ViewModel()
+        {
+            Data = new ObservableCollection<Person>()
+            {
+                new Person { Name = "David", Height = 180 },
+                new Person { Name = "Michael", Height = 170 },
+                new Person { Name = "Steve", Height = 160 },
+                new Person { Name = "Joel", Height = 182 }
+            };
+        }
+    }
+}
