--- SfChart_GettingStarted/MainPage.xaml
@@ -1,45 +1,0 @@
-﻿<Page
-    x:Class="SfChart_GettingStarted.MainPage"
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:SfChart_GettingStarted"
-    xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:syncfusion="using:Syncfusion.UI.Xaml.Charts"
-    xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    mc:Ignorable="d"
-    Background="{ThemeResource ApplicationPageBackgroundThemeBrush}" Height="350" Width="525">
-    <!--Setting DataContext-->
-    <Page.DataContext>
-        <local:ViewModel></local:ViewModel>
-    </Page.DataContext>
-
-    <Grid>
-        <syncfusion:SfChart Header="Chart">
-            <!--Initialize the horizontal axis for SfChart-->
-            <syncfusion:SfChart.PrimaryAxis>
-                <syncfusion:CategoryAxis Header="Name" />
-            </syncfusion:SfChart.PrimaryAxis>
-            <syncfusion:SfChart.SecondaryAxis>
-                <syncfusion:NumericalAxis Header="Height(in cm)"/>
-            </syncfusion:SfChart.SecondaryAxis>
-
-            <!--Adding Legend to the SfChart-->
-            <syncfusion:SfChart.Legend>
-                <syncfusion:ChartLegend />
-            </syncfusion:SfChart.Legend>
-
-            <!--Initialize the series for SfChart-->
-            <syncfusion:ColumnSeries  ItemsSource="{Binding Data}"
-                                      XBindingPath="Name"
-                                      YBindingPath="Height"
-                                      Label="Heights"
-                                      ShowTooltip="True">  
-                <!--Add data marker to the series-->
-                    <syncfusion:ColumnSeries.DataMarker>
-                        <syncfusion:ChartDataMarker DataMarkerPosition="Top" LabelPosition="Inner"  ShowLabel="True" />
-                    </syncfusion:ColumnSeries.DataMarker>
-            </syncfusion:ColumnSeries>          
-
-        </syncfusion:SfChart>
-    </Grid>
-</Page>
