--- SfChart_GettingStarted/Model/Model.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SfChart_GettingStarted
-{
-    public class Person
-    {
-        public string Name { get; set; }
-
-        public double Height { get; set; }
-    }
-}
