--- SfChart_GettingStarted/SfChart_GettingStarted.sln
@@ -1,35 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30317.65
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SfChart_GettingStarted", "SfChart_GettingStarted.csproj", "{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Debug|x64.ActiveCfg = Debug|x64
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Debug|x64.Build.0 = Debug|x64
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Debug|x64.Deploy.0 = Debug|x64
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Debug|x86.ActiveCfg = Debug|x86
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Debug|x86.Build.0 = Debug|x86
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Debug|x86.Deploy.0 = Debug|x86
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Release|x64.ActiveCfg = Release|x64
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Release|x64.Build.0 = Release|x64
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Release|x64.Deploy.0 = Release|x64
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Release|x86.ActiveCfg = Release|x86
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Release|x86.Build.0 = Release|x86
-		{789D5FE1-31FA-4F8E-A4EC-D3CF10EFDDDC}.Release|x86.Deploy.0 = Release|x86
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {10B0A038-AF6E-40DA-BDCD-B9930FE9BD77}
-	EndGlobalSection
-EndGlobal
