--- SfChart_GettingStarted/ViewModel/ViewModel.cs
@@ -1,24 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SfChart_GettingStarted
-{
-    public class ViewModel
-    {
-        public List<Person> Data { get; set; }
-
-        public ViewModel()
-        {
-            Data = new List<Person>()
-            {
-                new Person { Name = "David", Height = 180 },
-                new Person { Name = "Michael", Height = 170 },
-                new Person { Name = "Steve", Height = 160 },
-                new Person { Name = "Joel", Height = 182 }
-            };
-        }
-    }
-}
